Changed wall initialization
Changed the way walls are initialized and some more. Current build does not work, when clicking play unity will stop responding.

+++ Assets/Resources/Scripts/Classes/Actor/ExtendedActors/Chicken/ChickenSpawnerScript.cs
@@ -10,8 +10,8 @@ using TileSpace;
 public class ChickenSpawnerScript : CreateSpawnerScript
 {
 
-    public override void CreateSpawner(Structure Structure, int X, int Y)
+    public override void CreateSpawner(Structure Structure, float X, float Y)
     {
-        new ChickenSpawner(Structure.TileSpawnerArray[X,Y]);
+        new ChickenSpawner(Structure.TileSpawnerArray[(int)X,(int)Y]);
     }
 }
