+++ Assets/Resources/Scripts/StructureScripts/StructurePropertiesScript.cs
@@ -31,6 +31,6 @@ public class StructurePropertiesScript : MonoBehaviour
 
 public class CreateSpawnerScript : MonoBehaviour
 {
-    public virtual void CreateSpawner(Structure Structure, int x, int y)
+    public virtual void CreateSpawner(Structure Structure, float x, float y)
     { }
 }
